+++ baldwin/email/__init__.py
@@ -2,11 +2,12 @@
 
 from baldwin.exceptions import EmailDeliveryError, EmailFetchError, EmailServiceError
 
-from .email_service import Email, EmailService
+from .email_service import DEFAULT_IMAP_FOLDER, Email, EmailService, MailboxFolders
 from .postgres_store import PostgresEmailVectorStore, StoreResult
 from .vectorization import EmailNormalizer, HashingVectorizer, NormalizedEmail
 
 __all__ = [
+    "DEFAULT_IMAP_FOLDER",
     "Email",
     "EmailDeliveryError",
     "EmailFetchError",
@@ -14,6 +15,7 @@ __all__ = [
     "EmailService",
     "EmailServiceError",
     "HashingVectorizer",
+    "MailboxFolders",
     "NormalizedEmail",
     "PostgresEmailVectorStore",
     "StoreResult",